Pull out shared functions from `etl presence-absence`
